englob Data  User

+++ app/Http/Controllers/API/admin/UserController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\API\Admin;
 use App\Http\Controllers\API\BaseController;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UsersRequest;
+use App\Http\Resources\UserResource;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class UserController extends  BaseController
 {
@@ -20,7 +22,9 @@ class UserController extends  BaseController
     {
         $users=User::with(['roles','permissions'])->get();
 
-        return $this->sendResponse($users,"Getting  Your Users Successfully  !");
+        return $this->sendResponse(
+            UserResource::collection(User::with(['roles','permissions'])->paginate(20))->response()->getData(true)
+            ,"Getting  Your Users Successfully  !");
 
 
     }
@@ -41,7 +45,7 @@ class UserController extends  BaseController
         $user->roles()->attach($user_role);
 
 
-        return $this->sendResponse( $user ,'User Added successfully !');
+        return $this->sendResponse(UserResource::make($user) ,'User Added successfully !');
     }
 
     /**
@@ -53,7 +57,7 @@ class UserController extends  BaseController
     public function show($id)
     {
         $user=User::with(['roles'])->find($id);
-        return $this->sendResponse( $user ,'User showed successfully !');
+        return $this->sendResponse(UserResource::make($user) ,"user retrieved successfully");
     }
 
     /**
@@ -64,11 +68,13 @@ class UserController extends  BaseController
      * @return \Illuminate\Http\Response
      */
     public function update(UsersRequest $request, $id)
-    {        
+    {
+
+     abort_unless(auth()->user->hasRole('admin'),Response::HTTP_UNAUTHORIZED,'You are Not Allowed to destroy this User');
         $validatedUser = $request->validated();
         $user=User::findOrfail($id);
         $user->update($validatedUser);
-        return $this->sendResponse( $user ,'User showed successfully !');
+        return $this->sendResponse(UserResource::make($user) ,'User showed successfully !');
     }
 
     /**
@@ -79,7 +85,9 @@ class UserController extends  BaseController
      */
     public function destroy($id)
     {
+        abort_unless(auth()->user->hasRole('admin'),Response::HTTP_UNAUTHORIZED,'You are Not Allowed to destroy this User');
         $user=User::findOrfail($id);
+        $user->deleteProfilePhoto();
         $user->delete();
         return response()->noContent();
     }
